requirements.txt
# requirements.txt
Django==5.0.6
python-decouple==3.8
psycopg2-binary==2.9.9
django-allauth==0.61.1